backend/requirements.txt
fastapi
uvicorn
pydantic
pydantic-settings
sqlalchemy
sqlmodel
google-genai
google-generativeai
langchain-google-genai
python-jose[cryptography]
passlib[bcrypt]
python-multipart
python-dotenv
psycopg2-binary
pgvector
reportlab
